+++ src/react/index.ts
@@ -1,0 +1,1 @@
+export { ManageBillingButton } from "./ManageBillingButton.js";